posts and comments works, not secured

+++ src/posts/posts.controller.js
@@ -6,19 +6,21 @@ import cloudinary from 'cloudinary'
 
 export const createPost = async (req, res, next) => {
     const newPost = new Post(req.body);
+
     try {
         const cloudinaryResult = await uploadImage(req.file.buffer);
+        newPost.user = req.payload.id
         newPost.img = cloudinaryResult.secure_url;
         newPost.cloudinaryId = cloudinaryResult.public_id;
         const savedPost = await newPost.save();
         if (!savedPost) {
             res.status(400).json({message:'Post not saved! Try again.'});
         }
+
         // * Funktion um den Post direkt in den User zu pushen 
-        const user = await User.findById(req.userId);
+        const user = await User.findById(newPost.user);
         if (!user) {
-          return res.status(404).json({ message: 'User not found!' }); // ! !!!!FEHLER IN DER KONSOLE GEWOLLT!!!!
-          // * diese Funktion geht erst wenn wir die routen mit dem Token schützen und die ID des users hier erhalten. 
+          return res.status(404).json({ message: 'User not found!' });
         }
         user.posts.push(savedPost._id);
         await user.save();
@@ -52,14 +54,25 @@ export const updatePost = async (req, res, next) => {
   
   export const deletePost = async (req, res, next) => {
     const postId = req.params.id;
+  
     try {
       const post = await Post.findById(postId);
       if (!post) {
         return res.status(404).json({ message: 'Post not found!' });
       }
-      await cloudinary.v2.uploader.destroy(post.cloudinaryId);
+  
+      const user = await User.findById(post.user);
+      if (!user) {
+        return res.status(404).json({ message: 'User not found!' });
+      }
+  
+      user.posts = user.posts.filter(userPostId => userPostId.toString() !== postId);
+  
+      await user.save();
+  
       await Post.findByIdAndDelete(postId);
-      res.status(200).json({ message: 'Post sucessfully deleted!' });
+  
+      res.status(200).json({ message: 'Post successfully deleted!' });
     } catch (err) {
       next(err);
     }
@@ -68,7 +81,8 @@ export const updatePost = async (req, res, next) => {
   export const getPost = async (req, res, next) => {
     const postId = req.params.id
     try {
-      const post = await Post.findById(postId)
+      const post = await Post.findById(postId).populate('comments');
+      console.log(post);
       res.status(200).json(post)
     } catch (err) {
       next(err);
